Update from original head fork

+++ src/Workspaces/CSharp/Portable/CodeGeneration/ExpressionGenerator.cs
@@ -114,7 +114,7 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeGeneration
             }
             else if (value is byte)
             {
-                return GenerateLiteralExpression(type, (byte)value, LiteralSpecialValues.ByteSpecialValues, null, canUseFieldReference, (s, v) => SyntaxFactory.Literal(s, (int)v));
+                return GenerateLiteralExpression(type, (byte)value, LiteralSpecialValues.ByteSpecialValues, null, canUseFieldReference, (s, v) => SyntaxFactory.Literal(s, v));
             }
             else if (value is ushort)
             {
@@ -138,7 +138,7 @@ namespace Microsoft.CodeAnalysis.CSharp.CodeGeneration
             }
             else if (value is decimal)
             {
-                return GenerateLiteralExpression(type, (decimal)value,  LiteralSpecialValues.DecimalSpecialValues, null, canUseFieldReference, SyntaxFactory.Literal);
+                return GenerateLiteralExpression(type, (decimal)value, LiteralSpecialValues.DecimalSpecialValues, null, canUseFieldReference, SyntaxFactory.Literal);
             }
             else if (type == null || type.IsReferenceType || type.IsPointerType())
             {
